refactor(audit): extract sync lifecycle callbacks into a submodule
Phase 4 of the structure audit. Splits the 3157-line sync/lifecycle.rs by
moving the progress-callback block (TransferContext, handle_transfer_progress,
the build_*_callback builders, intent spawns, setup_progress_handlers) verbatim
into a new sync/lifecycle/callbacks.rs. lifecycle.rs drops to ~2530 lines and
re-exports build_file_synced_callback (pub) + setup_progress_handlers
(pub(crate)) so all call paths are unchanged.

Pure module move. Updated the hippius_activity_truth static-shape test to read
the new file location. Verified: clippy -D warnings + fmt clean; full suite
(incl. lifecycle + activity-truth tests) green.
